refactor(contextUtil): remove as no longer necessary
